change type

+++ src/main/java/com/nc13/StayHo/model/mapper/ImgMapper.java
@@ -19,7 +19,7 @@ public interface ImgMapper {
 
     Integer deleteReview(@Param("reviewId") int reviewId);
 
-    List<?> selectRoom(@Param("id") Long id);
+    List<?> selectRoom(@Param("id") int id);
 
     List<?> selectReview(@Param("id") int id);
