+++ src/main/java/com/nc13/StayHo/model/model/RoomDescriptionDTO.java
@@ -4,14 +4,14 @@ import lombok.Data;
 
 @Data
 public class RoomDescriptionDTO {
-    private Long roomId;
+    private int roomId;
     private boolean bath;
     private int bed;
     private String view;
     private String content;
 
     public RoomDescriptionDTO(){}
-    public RoomDescriptionDTO(Long roomId, boolean bath, int bed, String view, String content){
+    public RoomDescriptionDTO(int roomId, boolean bath, int bed, String view, String content){
         this.roomId=roomId;
         this.bath=bath;
         this.bed= bed;
